snippin_tools/requirements.txt
numpy
opencv-python
Pillow
mss
flet
screeninfo